refactor attention backend into FA2 and xFormers
