+++ .gitignore
@@ -196,3 +196,6 @@ _build/
 # HIP files generated by PyTorch
 *.hip
 *_hip*
+
+# Third-party tools
+aphrodite/thirdparty_files/
