+++ aphrodite/modeling/layers/attention/__init__.py
@@ -1,0 +1,5 @@
+from aphrodite.modeling.layers.attention.attention import Attention
+
+__all__ = [
+    "Attention",
+]
