+++ aphrodite/modeling/models/mixtral.py
@@ -30,7 +30,7 @@ from transformers import MixtralConfig
 
 from aphrodite.common.config import LoRAConfig
 from aphrodite.modeling.metadata import InputMetadata
-from aphrodite.modeling.layers.attention import PagedAttention
+from aphrodite.modeling.layers.attention import Attention
 from aphrodite.modeling.layers.triton_kernel.fused_moe import fused_moe
 from aphrodite.modeling.layers.layernorm import RMSNorm
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
@@ -218,7 +218,7 @@ class MixtralAttention(nn.Module):
             base=int(self.rope_theta),
             is_neox_style=is_neox_style,
         )
-        self.attn = PagedAttention(
+        self.attn = Attention(
             self.num_heads,
             self.head_dim,
             self.scaling,
